+++ tests/test_nb_simple.py
@@ -5,10 +5,10 @@ sys.path.insert(0, str(ROOT / 'build'))
 
 MODEL = ROOT / 'models' / 'gemma-2-9b-it-IQ4_XS.gguf'
 
-import pbllama as pb
+import nbllama as nb
 
-def test_simple():
-    params = pb.gpt_params()
+def test_nb_simple():
+    params = nb.gpt_params()
     params.prompt = "Hello my name is"
     params.n_predict = 32
 
@@ -18,19 +18,19 @@ def test_simple():
 
     # init LLM
 
-    pb.llama_backend_init()
-    pb.llama_numa_init(params.numa)
+    nb.llama_backend_init()
+    nb.llama_numa_init(params.numa)
 
     # initialize the model
 
     # llama_model_params model_params = llama_model_params_from_gpt_params(params)
-    model_params = pb.llama_model_params_from_gpt_params(params)
+    model_params = nb.llama_model_params_from_gpt_params(params)
 
     # set local test model
     params.model = str(MODEL)
 
     # llama_model * model = llama_load_model_from_file(params.model, model_params)
-    model = pb.llama_load_model_from_file(params.model, model_params)
+    model = nb.llama_load_model_from_file(params.model, model_params)
 
     # if (model == NULL) {
     #     fprintf(stderr , "%s: error: unable to load model\n" , __func__)
@@ -40,10 +40,10 @@ def test_simple():
     # initialize the context
 
     # llama_context_params ctx_params = llama_context_params_from_gpt_params(params)
-    ctx_params = pb.llama_context_params_from_gpt_params(params)
+    ctx_params = nb.llama_context_params_from_gpt_params(params)
 
     # llama_context * ctx = llama_new_context_with_model(model, ctx_params)
-    ctx = pb.llama_new_context_with_model(model, ctx_params)
+    # ctx = nb.llama_new_context_with_model(model, ctx_params)
 
     # if (ctx == NULL) {
     #     fprintf(stderr , "%s: error: failed to create the llama_context\n" , __func__)
@@ -54,7 +54,7 @@ def test_simple():
 
     # std::vector<llama_token> tokens_list
     # tokens_list = ::llama_tokenize(ctx, params.prompt, true)
-    # tokens_list = pb.llama_tokenize(ctx, params.prompt, True) # CRASH!!
+    # tokens_list = nb.llama_tokenize(ctx, params.prompt, True) # CRASH!!
 
     # const int n_ctx    = llama_n_ctx(ctx)
     # const int n_kv_req = tokens_list.size() + (n_predict - tokens_list.size())